+++ ThirdPartyAdapters/facebook/CHANGELOG.md
@@ -1,10 +1,14 @@
 ## Facebook Android Mediation Adapter Changelog
 
-#### Next Version
+#### 5.9.0.2
 - Added support for rewarded interstitial ads.
 - Updated the adapter to support inline adaptive banner requests.
 - Fixed an issue where open-bidding banner ads always render full-width.
 - Updated the minimum required Google Mobile Ads SDK version to 19.2.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 19.2.0.
+- Facebook SDK version 5.9.0.
 
 #### 5.9.0.1
 - Adapter now forwards an error if the FAN SDK encounters an error while presenting an interstitial/rewarded ad.